+++ docs/changelog.rst
@@ -9,8 +9,13 @@ CHANGELOG
 
 - Improve information about upgrading geotrek-admin version with debian
 
+**Improvements**
+
+- Change infrastructure condition field to ManyToMany field (#3970)
+
 **Bug fixes**
 - Fix view `v_treks` (fixes #4099)
+
 
 2.106.0 (2024-05-15)
 --------------------
@@ -62,7 +67,6 @@ CHANGELOG
 - Homogenize `structure` field in APIv2 to always return ids instead of names (#3007)
 - Add structures to Aggregator (#3569)
 - Allow to filter flatpages by portal on admin list page
-
 
 2.104.2 (2024-04-04)
 --------------------